web_app/backend/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
python-multipart==0.0.20
pydantic==2.12.3
sqlalchemy==2.0.44
aiohttp==3.13.1
python-dotenv==1.1.1